refactor(theta): consolidate binomial bounds
